Rework packetization section
In line with our established principles for retransmission, I've reworked the
description of packetization.  The description now concentrates on the
information that is being repaired in response to perceived loss.  This should
help avoid the confusion about retransmission.

Two new subsections are added to the packetization section.  One covers the
processing of packets and includes the existing text on processing requirements
before acknowledgment.  The other includes the retransmission logic.

The retransmission section still mentions frame types, but I've tried to make
that secondary to the description of the information that is being repaired.

I also removed mention of "Regular QUIC packets", which only occurred in 3
places.

Closes #463, #765.